requirement.txt
absl-py==2.0.0
astor==0.8.1
certifi==2022.12.7
charset-normalizer==3.3.2
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.59.3
h5py==2.10.0
idna==3.6
importlib-metadata==6.7.0
joblib==1.3.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.1
numpy==1.21.6
opencv-python==4.5.5.64
openpyxl==3.1.2
packaging==23.2
Pillow==9.5.0
pip==22.3.1
protobuf==3.20.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==0.23.0
scipy==1.7.3
setuptools==65.6.3
six==1.16.0
tensorboard==1.14.0
tensorboardX==2.6.2.2
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==2.3.0
threadpoolctl==3.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.16.0
zipp==3.15.0
